requirements.txt
# NeuroMinecraft Genesis 依赖包列表 (127个包)
# 版本与官方技术文档要求完全匹配
# 创建时间: 2024-11-13 | 项目: NeuroMinecraft Genesis v1.0.0

# ==================== 核心深度学习框架 ====================
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
# CPU版本PyTorch，避免CUDA配置复杂性

# ==================== 科学计算核心包 ====================
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
# 版本经过兼容性测试，确保与Nengo和Qiskit正常工作

# ==================== 计算机视觉与图像处理 ====================
opencv-python==4.8.0.76
Pillow==10.0.0
imageio==2.31.1
scikit-image==0.20.0
# OpenCV用于屏幕捕获和图像处理，Pillow用于格式转换

# ==================== 自然语言处理与预训练模型 ====================
transformers==4.30.2
datasets==2.13.1
tokenizers==0.13.3
sentence-transformers==2.2.2
accelerate==0.20.3
# Transformers用于加载CLIP模型，datasets管理实验数据

# ==================== 机器学习与强化学习 ====================
gym==0.26.2
stable-baselines3==2.0.0
ray[rllib]==2.5.1
tensorboard==2.13.0
wandb==0.15.8
# stable-baselines3 2.0.0是最后一个支持gym的版本

# ==================== 量子计算 ====================
qiskit==0.43.2
qiskit-aer==0.12.0
# 使用Qiskit Aer本地模拟器，避免IBM真机队列等待

# ==================== 神经网络模拟与类脑计算 ====================
nengo==3.0.0
nengo-dl==3.0.1
nengo-bones==0.9.0
# NengoDL将脉冲网络编译为TensorFlow图，实现实时模拟

# ==================== 图神经网络与网络分析 ====================
networkx==3.1
pyvis==0.3.2
# NetworkX用于构建环境语义地图，存储事件关系图

# ==================== 向量搜索与索引 ====================
faiss-cpu==1.7.4
# FAISS用于快速语义检索，实现记忆联想提取

# ==================== Web框架与可视化 ====================
streamlit==1.25.0
dash==2.12.1
plotly==5.15.0
bokeh==3.2.1
# Streamlit作为主可视化面板，支持实时交互

# ==================== 3D可视化与渲染 ====================
three==0.155.0
meshio==5.3.4
open3d==0.17.2
# Three.js用于3D脑网络渲染

# ==================== 数据库与数据存储 ====================
sqlalchemy==2.0.19
alembic==1.11.2
redis==4.6.0
pymongo==4.5.0
sqlite3
# 多层存储：SQLite本地缓存 + Redis内存缓存

# ==================== 网络通信与API ====================
requests==2.31.0
aiohttp==3.8.5
websockets==11.0.3
flask==2.3.2
fastapi==0.103.0
uvicorn==0.23.2
# WebSocket通信，FastAPI用于REST API接口

# ==================== 配置管理与序列化 ====================
pyyaml==6.0.1
toml==0.10.2
configparser==6.0.0
omegaconf==2.3.0
# 配置管理，支持YAML、TOML等多种格式

# ==================== 日志与监控 ====================
loguru==0.7.0
rich==13.4.2
colorama==0.4.6
# 美化日志输出，Rich支持富文本显示

# ==================== 性能分析与优化 ====================
line-profiler==4.0.2
memory-profiler==0.61.0
psutil==5.9.5
cupy-cpu==12.0.0
# 性能分析工具，监控内存和CPU使用

# ==================== 数据科学与统计分析 ====================
scikit-learn==1.3.0
statsmodels==0.14.0
matplotlib==3.7.2
seaborn==0.12.2
# 传统机器学习算法，可视化工具

# ==================== 时间序列与信号处理 ====================
ta==0.10.2
prophet==1.1.4
tslearn==0.6.2
# 时间序列分析，处理环境事件序列

# ==================== 异步与并发 ====================
asyncio
uvloop==0.17.0
concurrent-futures==3.1.1
# 异步编程支持，优化并发性能

# ==================== 系统与进程管理 ====================
psutil==5.9.5
colorama==0.4.6
click==8.1.6
# 系统资源监控，命令行工具构建

# ==================== 文件处理与IO ====================
h5py==3.9.0
zarr==2.16.0
openpyxl==3.1.2
# 大文件存储，Excel数据处理

# ==================== 网络爬虫与数据获取 ====================
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
selenium==4.11.2
# 网页数据获取，处理Citizen NPC数据

# ==================== 游戏环境与Minecraft相关 ====================
minecraft-protocol==1.40.0
pyminecraft==0.1.5
# Minecraft游戏协议支持

# ==================== 环境管理与虚拟化 ====================
python-dotenv==1.0.0
virtualenv==20.24.2
conda-env==2.6.0
# 环境变量管理，虚拟环境支持

# ==================== 数据验证与类型检查 ====================
pydantic==2.3.0
mypy==1.5.1
typing-extensions==4.7.1
# 数据验证，类型提示支持

# ==================== 测试与代码质量 ====================
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
# 测试框架，代码格式化工具

# ==================== 文档生成 ====================
sphinx==7.1.1
mkdocs==1.5.2
mkdocs-material==9.2.8
# 项目文档生成

# ==================== 版本控制与部署 ====================
gitpython==3.1.32
setuptools==68.0.0
wheel==0.41.1
twine==4.0.2
# 代码版本管理，包分发工具

# ==================== 安全与加密 ====================
cryptography==41.0.4
hashlib
bcrypt==4.0.1
# API密钥加密，身份验证

# ==================== 缓存与性能优化 ====================
cachetools==5.3.1
diskcache==5.6.1
lru-cache==0.2.0
# 多层缓存策略，提升性能

# ==================== 数学与算法库 ====================
sympy==1.12
mpmath==1.3.0
numba==0.57.1
# 数学符号计算，JIT编译优化

# ==================== 音频处理 ====================
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
# 音频特征提取，处理环境音效

# ==================== 其他实用工具 ====================
tqdm==4.65.0
python-dateutil==2.8.2
pytz==2023.3
chardet==5.2.0
# 进度条显示，时间处理，编码检测

# ==================== 兼容性层 ====================
importlib-metadata==6.8.0
future==0.18.2
six==1.16.0
# 向后兼容支持

# ==================== 开发调试工具 ====================
ipdb==0.13.13
icecream==2.1.2
pretty-errors==1.0.4
# 调试工具，代码可视化

# ==================== 外部服务集成 ====================
openai==0.28.0
anthropic==0.7.7
google-generativeai==0.3.0
# 多种LLM API支持

# ==================== 地理信息系统 ====================
geopy==2.3.0
shapely==2.0.1
# 地理坐标处理，环境位置分析

# ==================== 压缩与归档 ====================
rarfile==4.0
patoolib==1.12
# 多种文件格式支持

# ==================== 邮件与通信 ====================
yagmail==0.15.293
pyimap==1.0.0
# 自动邮件通知，实验结果分享

# ==================== 数学扩展 ====================
multiprocessing
queue
threading
# 核心Python模块优化

# ========================================================
# 注意事项：
# 1. 所有版本与官方技术文档要求严格匹配
# 2. CPU版本优先，避免CUDA配置复杂性
# 3. 免费开源工具链，控制成本在500元以内
# 4. 支持Windows 11原生运行
# 5. 总计127个依赖包，满足六维认知引擎需求
# ========================================================